avoid too many downloads

+++ tests/conftest.py
@@ -3,6 +3,7 @@ import gc
 import json
 import os
 import sys
+import tempfile
 from collections import UserList
 from enum import Enum
 from typing import (Any, Callable, Dict, List, Optional, Tuple, TypedDict,
@@ -761,17 +762,24 @@ def num_gpus_available():
     return cuda_device_count_stateless()
 
 
+temp_dir = tempfile.gettempdir()
+_dummy_path = os.path.join(temp_dir, "dummy_opt")
+
+
 @pytest.fixture
 def dummy_opt_path():
-    opt_path = snapshot_download(repo_id="facebook/opt-125m",
-                                 ignore_patterns=[
-                                     "*.bin", "*.bin.index.json", "*.pt",
-                                     "*.h5", "*.msgpack"
-                                 ])
-    json_path = os.path.join(opt_path, "config.json")
-    with open(json_path, "r") as f:
-        config = json.load(f)
-    config["architectures"] = ["MyOPTForCausalLM"]
-    with open(json_path, "w") as f:
-        json.dump(config, f)
-    return opt_path
+    json_path = os.path.join(_dummy_path, "config.json")
+    if not os.path.exists(_dummy_path):
+        snapshot_download(repo_id="facebook/opt-125m",
+                          local_dir=_dummy_path,
+                          ignore_patterns=[
+                              "*.bin", "*.bin.index.json", "*.pt", "*.h5",
+                              "*.msgpack"
+                          ])
+        assert os.path.exists(json_path)
+        with open(json_path, "r") as f:
+            config = json.load(f)
+        config["architectures"] = ["MyOPTForCausalLM"]
+        with open(json_path, "w") as f:
+            json.dump(config, f)
+    return _dummy_path
